Staging: hv: storvsc: Cleanup storvsc_device_alloc()
The code in storvsc_device_alloc() is not needed as this would be
done by default. Get rid of it. We still keep the function as we use
this hook to allocate per-LUN memory pools in a later patch.

Signed-off-by: K. Y. Srinivasan <kys@microsoft.com>
Signed-off-by: Haiyang Zhang <haiyangz@microsoft.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
(cherry picked from commit 2781866484624fabbe373b95cfd20d44d2c9a69e)

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>