vga_switcheroo: Add DDC lock for LVDS detection.
Some GPU drivers don't support hotplugging LVDS displays and disable the LVDS
output completely if no display is connected to it at boot. But during boot of a
system with switchable graphics the internal display is only connected to one
card. This adds a function that switches the DDC line of the internal display
so the LVDS panel can be detected by the secondary GPU driver.